beginning of scroll attempt
