+++ package.json
@@ -15,8 +15,10 @@
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-router-dom": "^5.1.2",
+    "react-section": "^1.1.11",
     "react-transition-group": "^4.3.0",
-    "react-undraw-illustrations": "^2.0.3"
+    "react-undraw-illustrations": "^2.0.3",
+    "react-scroll": "^1.7.10"
   },
   "devDependencies": {
     "@babel/cli": "^7.7.4",
