+++ src/components/JobsCard.jsx
